feat: author and resolve pending insertion proposals
