+++ packages/demo/e2e/main.tsx
@@ -1,4 +1,12 @@
 import { createRoot } from "react-dom/client";
 import ReviewEditorFixture from "./ReviewEditorFixture";
 
-createRoot(document.getElementById("root")!).render(<ReviewEditorFixture />);
+import { InsertionFixture } from "./InsertionFixture";
+
+createRoot(document.getElementById("root")!).render(
+  location.search === "?insertions" ? (
+    <InsertionFixture />
+  ) : (
+    <ReviewEditorFixture />
+  ),
+);
