+++ packages/demo/package.json
@@ -16,6 +16,7 @@
     "react-dom": "^19.2.3"
   },
   "devDependencies": {
+    "@lexical/history": "0.49.0",
     "@tailwindcss/postcss": "^4.1.18",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
